feat: money as exact decimal strings, Snowflake ids as JSON strings (#39)
Storage and engine were already 8-dp fixed-point longs; the double leak was
purely the API boundary. Now:
- REST/WS/gRPC emit money as exact decimal strings (FixedPoint.format) and
  omsOrderId/clusterOrderId as JSON strings (2^53 JS precision, ui#25).
- Input: decimal strings are canonical and parsed exactly
  (FixedPoint.parse); legacy JSON numbers still accepted one release
  (deprecated, fromDouble path) so pre-freeze clients keep working.
- CreateOrderRequest/OrderResponse carry fixed-point longs internally
  (FixedPointJson Jackson codec); updateOrder takes fixed-point longs;
  balances/deposit/withdraw round-trip strings; proto money double→string
  (ids stay int64: protobuf is 64-bit native); GrpcAccountService reads the
  now-string balance map.
- Frozen error shape: {"error", "code"} with codes VALIDATION/UNAUTHORIZED/
  FORBIDDEN/NOT_FOUND/REJECTED/ADMIN_UNAVAILABLE/INTERNAL (401 handler
  included); malformed money strings die 400 VALIDATION, not 500.

WireContractTest pins the shapes (id strings, 8-dp money strings, exact
string parse, legacy number tolerance, malformed rejection). Full suite
green. Requires match-common with FixedPoint.parse (match#63, merged).

+++ oms-api/src/main/java/com/openexchange/oms/api/OrderService.java
@@ -15,7 +15,8 @@ public interface OrderService {
 
     CancelOrderResponse cancelOrder(long omsOrderId);
 
-    Map<String, Object> updateOrder(long omsOrderId, double newPrice, double newQuantity);
+    /** Prices/quantities are 8-dp fixed-point longs; 0 = keep the current value. */
+    Map<String, Object> updateOrder(long omsOrderId, long newPrice, long newQuantity);
 
     OrderResponse getOrder(long omsOrderId);
